requirements.txt
pytest
selenium
webdriver-manager
python-dotenv
requests
allure-pytest